+++ api/src/main/java/ca/bc/gov/educ/penreg/api/repository/PenRequestBatchRepository.java
@@ -1,6 +1,7 @@
 package ca.bc.gov.educ.penreg.api.repository;
 
 import ca.bc.gov.educ.penreg.api.model.v1.PenRequestBatchEntity;
+import ca.bc.gov.educ.penreg.api.model.v1.PenRequestBatchStudentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Query;
@@ -32,28 +33,6 @@ public interface PenRequestBatchRepository extends JpaRepository<PenRequestBatch
    * @return the list
    */
   List<PenRequestBatchEntity> findByPenRequestBatchStatusCode(String penRequestBatchStatusCode);
-
-  /**
-   * Find all pen numbers that were issued to more than one student in pen batch.
-   *
-   * @param  penRequestBatchID the pen request batch ID
-   * @return a list of PEN numbers that were assigned to more than one student
-   */
-  @Query(value = "select PEN_REQUEST_BATCH_STUDENT_ID\n"
-      + "from PEN_REQUEST_BATCH_STUDENT\n"
-      + "where STUDENT_ID in (\n"
-      + "SELECT\n"
-      + "    student_id\n"
-      + "FROM\n"
-      + "    PEN_REQUEST_BATCH_STUDENT\n"
-      + "WHERE PEN_REQUEST_BATCH_ID in :penRequestBatchIDs\n"
-      + "GROUP BY\n"
-      + "    student_id\n"
-      + "HAVING\n"
-      + "    COUNT( student_id ) > 1)\n"
-      + "AND PEN_REQUEST_BATCH_ID in :penRequestBatchIDs",
-    nativeQuery = true)
-  List<String> findSameAssignedPensByPenRequestBatchID(@Param("penRequestBatchIDs") List<UUID> penRequestBatchID);
 
   List<PenRequestBatchEntity> findByPenRequestBatchStatusCodeAndSchoolGroupCode(String penRequestBatchStatusCode, String schoolGroupCode);
