+++ api/src/main/java/ca/bc/gov/educ/penreg/api/service/PenRequestBatchService.java
@@ -188,17 +188,6 @@ public class PenRequestBatchService {
   }
 
   /**
-   * Find all of the same PEN numbers issued to more than one student within a list of pen request batch ids.
-   *
-   * @param penRequestBatchID the pen request batch id
-   * @return a list of pen batch request student ids that have the same assigned pen number.
-   */
-  @Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
-  public List<String> getAllSamePensWithinPenRequestBatchByID(final List<UUID> penRequestBatchID) {
-    return this.getRepository().findSameAssignedPensByPenRequestBatchID(penRequestBatchID);
-  }
-
-  /**
    * Create pen request batch pen request batch entity.
    *
    * @param penRequestBatchEntity the pen request batch entity
