+++ api/src/test/resources/mock_pen_req_batch_student_with_same_student_id.json
@@ -1,0 +1,88 @@
+[
+  {
+
+    "penRequestBatchStudentStatusCode": "LOADED",
+    "localID": "1488645",
+    "submittedPen": "123456789",
+    "legalLastName": "JOHNSTON",
+    "legalFirstName": "ANGEL",
+    "legalMiddleNames": "MARIA LYNN",
+    "usualLastName": "JEB",
+    "usualFirstName": "JEB",
+    "usualMiddleNames": "JEB",
+    "dob": "20060628",
+    "genderCode": "F",
+    "gradeCode": "02",
+    "postalCode": "Y1A4V1",
+    "bestMatchPEN":  "123456789",
+    "assignedPEN": "123456789",
+    "studentID": "c0a8001f-7f56-16ae-817f-56e70f100000"
+  },
+  {
+
+    "penRequestBatchStudentStatusCode": "LOADED",
+    "localID": "2046291",
+    "submittedPen": null,
+    "legalLastName": "BJORK",
+    "legalFirstName": "HAYLEE",
+    "legalMiddleNames": "ELIZABETH",
+    "usualLastName": "BJORK",
+    "usualFirstName": "HAYLEE",
+    "usualMiddleNames": null,
+    "dob": "20120418",
+    "genderCode": "F",
+    "gradeCode": "EL",
+    "postalCode": "V0J1P0",
+    "studentID": "c0a8001f-7f56-16ae-817f-56e70f100000"
+  },
+  {
+
+    "penRequestBatchStudentStatusCode": "LOADED",
+    "localID": "2046301",
+    "submittedPen": null,
+    "legalLastName": "JOSEPH",
+    "legalFirstName": "BRAYDON",
+    "legalMiddleNames": "JAMIESON",
+    "usualLastName": "JOSEPH",
+    "usualFirstName": "BRAYDON",
+    "usualMiddleNames": null,
+    "dob": "20111207",
+    "genderCode": "M",
+    "gradeCode": "EL",
+    "postalCode": "V0J1P0",
+    "studentID": "c0a8001f-7f56-16ae-817f-56e70f100000"
+  },
+  {
+
+    "penRequestBatchStudentStatusCode": "LOADED",
+    "localID": "2046293",
+    "submittedPen": null,
+    "legalLastName": "JEBSTER",
+    "legalFirstName": "KEIDEN",
+    "legalMiddleNames": null,
+    "usualLastName": "WEBSTER",
+    "usualFirstName": "KEIDEN",
+    "usualMiddleNames": null,
+    "dob": "20121204",
+    "genderCode": "M",
+    "gradeCode": "EL",
+    "postalCode": "V0J1P0",
+    "studentID": "c0a8001f-7f56-16ae-817f-56e70f100000"
+  },
+  {
+    "penRequestBatchStudentStatusCode": "LOADED",
+    "localID": "221022",
+    "submittedPen": null,
+    "legalLastName": "JEY",
+    "legalFirstName": "JES",
+    "legalMiddleNames": "STU",
+    "usualLastName": "WEY",
+    "usualFirstName": "JES",
+    "usualMiddleNames": null,
+    "dob": "19901206",
+    "genderCode": "M",
+    "gradeCode": "GR",
+    "postalCode": "V8G1A7",
+    "studentID": "c0a8001f-7f56-16ae-817f-56e70f100000"
+  }
+]
